+++ lib/mapbox_maps_flutter.dart
@@ -88,3 +88,4 @@ part 'src/viewport/transitions/easing_viewport_transition.dart';
 part 'src/package_info.dart';
 part 'src/http/http_service.dart';
 part 'src/cancelable.dart';
+part 'src/deprecated.dart';