+++ lib/src/pigeons/point_annotation_messenger.dart
@@ -290,7 +290,6 @@ class PointAnnotation {
     this.iconHaloBlur,
     this.iconHaloColor,
     this.iconHaloWidth,
-    @Deprecated('Use [PointAnnotationManager.iconImageCrossFade] instead.')
     this.iconImageCrossFade,
     this.iconOcclusionOpacity,
     this.iconOpacity,
@@ -411,7 +410,7 @@ class PointAnnotation {
 
   /// Controls the transition progress between the image variants of icon-image. Zero means the first variant is used, one is the second, and in between they are blended together. Both images should be the same size and have the same type (either raster or vector).
   /// Default value: 0. Value range: [0, 1]
-  @Deprecated('Use [PointAnnotationManager.iconImageCrossFade] instead.')
+  /// Deprecated: Use `PointAnnotationManager.iconImageCrossFade` instead.
   double? iconImageCrossFade;
 
   /// The opacity at which the icon will be drawn in case of being depth occluded. Absent value means full occlusion against terrain only.
@@ -634,7 +633,6 @@ class PointAnnotationOptions {
     this.iconHaloBlur,
     this.iconHaloColor,
     this.iconHaloWidth,
-    @Deprecated('Use [PointAnnotationManager.iconImageCrossFade] instead.')
     this.iconImageCrossFade,
     this.iconOcclusionOpacity,
     this.iconOpacity,
@@ -752,7 +750,7 @@ class PointAnnotationOptions {
 
   /// Controls the transition progress between the image variants of icon-image. Zero means the first variant is used, one is the second, and in between they are blended together. Both images should be the same size and have the same type (either raster or vector).
   /// Default value: 0. Value range: [0, 1]
-  @Deprecated('Use [PointAnnotationManager.iconImageCrossFade] instead.')
+  /// Deprecated: Use `PointAnnotationManager.iconImageCrossFade` instead.
   double? iconImageCrossFade;
 
   /// The opacity at which the icon will be drawn in case of being depth occluded. Absent value means full occlusion against terrain only.
@@ -1072,51 +1070,6 @@ class PointAnnotationMessenger_PigeonCodec extends StandardMessageCodec {
         return PointAnnotationOptions.decode(readValue(buffer)!);
       default:
         return super.readValueOfType(type, buffer);
-    }
-  }
-}
-
-abstract class OnPointAnnotationClickListener {
-  static const MessageCodec<Object?> pigeonChannelCodec =
-      PointAnnotationMessenger_PigeonCodec();
-
-  void onPointAnnotationClick(PointAnnotation annotation);
-
-  static void setUp(
-    OnPointAnnotationClickListener? api, {
-    BinaryMessenger? binaryMessenger,
-    String messageChannelSuffix = '',
-  }) {
-    messageChannelSuffix =
-        messageChannelSuffix.isNotEmpty ? '.$messageChannelSuffix' : '';
-    {
-      final BasicMessageChannel<
-          Object?> pigeonVar_channel = BasicMessageChannel<
-              Object?>(
-          'dev.flutter.pigeon.mapbox_maps_flutter.OnPointAnnotationClickListener.onPointAnnotationClick$messageChannelSuffix',
-          pigeonChannelCodec,
-          binaryMessenger: binaryMessenger);
-      if (api == null) {
-        pigeonVar_channel.setMessageHandler(null);
-      } else {
-        pigeonVar_channel.setMessageHandler((Object? message) async {
-          assert(message != null,
-              'Argument for dev.flutter.pigeon.mapbox_maps_flutter.OnPointAnnotationClickListener.onPointAnnotationClick was null.');
-          final List<Object?> args = (message as List<Object?>?)!;
-          final PointAnnotation? arg_annotation = (args[0] as PointAnnotation?);
-          assert(arg_annotation != null,
-              'Argument for dev.flutter.pigeon.mapbox_maps_flutter.OnPointAnnotationClickListener.onPointAnnotationClick was null, expected non-null PointAnnotation.');
-          try {
-            api.onPointAnnotationClick(arg_annotation!);
-            return wrapResponse(empty: true);
-          } on PlatformException catch (e) {
-            return wrapResponse(error: e);
-          } catch (e) {
-            return wrapResponse(
-                error: PlatformException(code: 'error', message: e.toString()));
-          }
-        });
-      }
     }
   }
 }